third_party/blink/web_tests/editing/selection/triple-click-user-select-none-with-input.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>

function tripleClick(x, y) {
  return new Promise((resolve, reject) => {
    assert_own_property(window, 'chrome');
    assert_own_property(window.chrome, 'gpuBenchmarking');

    chrome.gpuBenchmarking.pointerActionSequence([{
      source: 'mouse',
      actions: [
        { name: 'pointerMove', x: x, y: y },
        { name: 'pointerDown', x: x, y: y },
        { name: 'pointerUp' },
        { name: 'pointerDown', x: x, y: y },
        { name: 'pointerUp' },
        { name: 'pointerDown', x: x, y: y },
        { name: 'pointerUp' },
      ],
    }], resolve);
  });
}

// The triple click should select the entire paragraph text, even when there's
// a user-select:none element at the start.
selection_test(
    [
        '<style>',
        '.no-select { user-select: none; }',
        '</style>',
        '<div id="target">',
        '<span class="no-select"><input type="range"></span>',
        '<span id="text">Please triple click in the middle of this text</span>',
        '</div>'
    ].join(''),
    async (selection) => {
        const elem = selection.document.getElementById('text');
        const x = selection.computeLeft(elem) + elem.offsetWidth / 2;
        const y = selection.computeTop(elem) + elem.offsetHeight / 2;
        await tripleClick(x, y);
    },
    [
        '<style>',
        '.no-select { user-select: none; }',
        '</style>',
        '<div id="target">',
        '<span class="no-select"><input type="range"></span>',
        '<span id="text">^Please triple click in the middle of this text|</span>',
        '</div>'
    ].join(''),
    'Triple click selects entire paragraph, skipping user-select:none input at start');

// Same test but with the input at the end of the paragraph.
selection_test(
    [
        '<style>',
        '.no-select { user-select: none; }',
        '</style>',
        '<div id="target">',
        '<span id="text">Please triple click in the middle of this text</span>',
        '<span class="no-select"><input type="range"></span>',
        '</div>'
    ].join(''),
    async (selection) => {
        const elem = selection.document.getElementById('text');
        const x = selection.computeLeft(elem) + elem.offsetWidth / 2;
        const y = selection.computeTop(elem) + elem.offsetHeight / 2;
        await tripleClick(x, y);
    },
    [
        '<style>',
        '.no-select { user-select: none; }',
        '</style>',
        '<div id="target">',
        '<span id="text">^Please triple click in the middle of this text|</span>',
        '<span class="no-select"><input type="range"></span>',
        '</div>'
    ].join(''),
    'Triple click selects entire paragraph, skipping user-select:none input at end');

// Test with non-selectable inputs at both ends of the paragraph.
selection_test(
    [
        '<style>',
        '.no-select { user-select: none; }',
        '</style>',
        '<div id="target">',
        '<span class="no-select"><input type="range"></span>',
        '<span id="text">Please triple click in the middle of this text</span>',
        '<span class="no-select"><input type="checkbox"></span>',
        '</div>'
    ].join(''),
    async (selection) => {
        const elem = selection.document.getElementById('text');
        const x = selection.computeLeft(elem) + elem.offsetWidth / 2;
        const y = selection.computeTop(elem) + elem.offsetHeight / 2;
        await tripleClick(x, y);
    },
    [
        '<style>',
        '.no-select { user-select: none; }',
        '</style>',
        '<div id="target">',
        '<span class="no-select"><input type="range"></span>',
        '<span id="text">^Please triple click in the middle of this text|</span>',
        '<span class="no-select"><input type="checkbox"></span>',
        '</div>'
    ].join(''),
    'Triple click selects entire paragraph, skipping user-select:none inputs at both ends');
</script>
